modify almost done

+++ src/components/NoteList.js
@@ -3,8 +3,9 @@ import { message, Menu, Modal } from 'antd';
 import { AppstoreOutlined, DeleteOutlined, ExclamationCircleOutlined } from '@ant-design/icons';
 import styled from 'styled-components';
 import axios from 'axios';
-
 import { DragDropContext, Droppable, Draggable } from "react-beautiful-dnd";
+
+const { SubMenu } = Menu;
 
 const getListStyle = isDraggingOver => ({
 	background: isDraggingOver ? "#7ec1ff" : "",
@@ -24,12 +25,14 @@ const getItemStyle = (isDragging, draggableStyle, active) => {
 		padding: "16px 8px",
 		borderBottom: "1px solid lightgray",
 		background: background,
+		borderRadius: "3px",
+		cursor: "default",
+
 		...draggableStyle
 	}
 };
 
 
-const { SubMenu } = Menu;
 
 const ContextMenu = styled.div`
 	display: none;
@@ -58,6 +61,18 @@ const MenuItemContent = styled.div`
 	}
 `;
 
+const DraggableItem = styled.div`
+	.title {
+		line-height: normal;
+		font-weight: bold;
+		font-size: 18px;
+	}
+	.date {
+		line-height: normal;
+		font-size: 14px;
+	}
+`;
+
 const ConfirmContent = styled.div`
 	text-align: center;
 	.title {
@@ -70,28 +85,23 @@ const ConfirmContent = styled.div`
 `;
 
 const NoteList = (props) => {
-	const { liveNoteList, setLiveNoteList, deletedNoteList, setDeletedNoteList, getNotes, updateNoteToServer } = props;
+	const { newId, liveNoteList, setLiveNoteList, deletedNoteList, setDeletedNoteList, getNotes, updateNoteToServer } = props;
 
 	const [curNote, setCurNote] = useState(null);
 
-	const fakeDeleteNote = (note) => {
-		let newNote = {
-			...note,
-			deleted: 1,
-		}
-		updateNoteToServer(newNote, getNotes);
-	}
-
 	const defaultNoteList = () => {
-		//delete server side empty note.
-		liveNoteList.forEach(item => {
-			if (item.title === 'New Note') {
-				axios.delete(`/note/${item.id}`).catch((err) => {
-					message.error('delete note error');
-					console.log(err);
-				});
-			}
-		});
+		if (newId == null) {
+			// newId is null represent that we already create a new note in DB, so we can delete it from DB
+			//delete server side empty note.
+			liveNoteList.forEach(item => {
+				if (item.title === 'New Note') {
+					axios.delete(`/note/${item.id}`).catch((err) => {
+						message.error('delete note error');
+						console.log(err);
+					});
+				}
+			});
+		}
 
 		setLiveNoteList((pre) => {
 			return pre.filter((item) => item.title !== 'New Note').map((item) => {
@@ -162,13 +172,15 @@ const NoteList = (props) => {
 							return pre.filter(item => item.title !== 'New Note');
 						});
 					} else {
-						fakeDeleteNote(curNote);
+						axios.delete(`/note/fake/${curNote.id}`).then(() => {
+							getNotes();
+						});
 					}
 					setCurNote(null);
 					document.getElementById("Menu").style.display = 'none';
 				}}
 			>
-				delete
+				move to trash
 			</div>
 		</ContextMenu>
 		<ContextMenu id='Menu2'>
@@ -204,15 +216,20 @@ const NoteList = (props) => {
 			>
 				delete
 			</div>
-			<div className='item' onClick={() => {
-				let newNote = {
-					...curNote,
-					deleted: 0,
-				}
-				updateNoteToServer(newNote, getNotes);
-				setCurNote(null);
-				document.getElementById("Menu2").style.display = 'none';
-			}}>recover</div>
+			<div
+				className='item'
+				onClick={() => {
+					let newNote = {
+						...curNote,
+						deleted: 0,
+					}
+					updateNoteToServer(newNote, getNotes);
+					setCurNote(null);
+					document.getElementById("Menu2").style.display = 'none';
+				}}
+			>
+				recover
+			</div>
 		</ContextMenu>
 		<Menu
 			defaultOpenKeys={['sub1']}
@@ -220,105 +237,81 @@ const NoteList = (props) => {
 			selectedKeys={getSelectedKeys()}
 		>
 			<SubMenu key="sub1" icon={<AppstoreOutlined />} title="Notes">
-				{
-					<DragDropContext
-						onDragStart={() => {
-							defaultNoteList();
-						}}
-						onDragEnd={(result) => {
-							if (!result.destination) {
-								return;
+				<DragDropContext
+					onDragStart={() => {
+						defaultNoteList();
+					}}
+					onDragEnd={(result) => {
+						if (!result.destination) {
+							return;
+						}
+
+						let ans = Array.from(liveNoteList);
+						const [removed] = ans.splice(result.source.index, 1);
+						ans.splice(result.destination.index, 0, removed);
+						for (let i = 0; i < ans.length; i++) {
+							ans[i].number = i;
+						}
+						setLiveNoteList(ans);
+
+						ans = ans.map(note => {
+							return {
+								...note,
+								content: JSON.stringify(note.content),
 							}
+						});
+						axios.put('/note/updateLiveNoteList', ans);
+					}}
+				>
+					<Droppable droppableId="droppable">
+						{(provided, snapshot) => {
+							return <div
+								{...provided.droppableProps}
+								ref={provided.innerRef}
+								style={getListStyle(snapshot.isDraggingOver)}
+							>
+								{liveNoteList.map((note, index) => {
+									return <Draggable key={note.id} draggableId={note.id} index={index}>
+										{(provided, snapshot) => (
+											<DraggableItem
+												ref={provided.innerRef}
+												{...provided.draggableProps}
+												{...provided.dragHandleProps}
+												style={getItemStyle(
+													snapshot.isDragging,
+													provided.draggableProps.style,
+													note.active
+												)}
+												onClick={() => {
+													handleClickLiveNote(note);
+												}}
+												onContextMenu={(e) => {
+													handleClickLiveNote(note);
+													setCurNote(note);
 
-							let ans = Array.from(liveNoteList);
-							const [removed] = ans.splice(result.source.index, 1);
-							ans.splice(result.destination.index, 0, removed);
-							for (let i = 0; i < ans.length; i++) {
-								ans[i].number = i;
-							}
-							setLiveNoteList(ans);
-
-							ans = ans.map(note => {
-								return {
-									...note,
-									content: JSON.stringify(note.content),
-								}
-							});
-							axios.put('/note/updateLiveNoteList', ans);
+													e.preventDefault();
+													const clickX = e.clientX;
+													const clickY = e.clientY;
+													const Menu = document.getElementById("Menu");
+													Menu.style.display = "block";
+													Menu.style.position = "absolute";
+													Menu.style.left = `${clickX}px`; Menu.style.top = `${clickY}px`;
+												}}
+											>
+												<div className='title'>{note.title}</div>
+												<div className='date'>{note.updateTime.format("yyyy/MM/DD HH:mm:ss")}</div>
+											</DraggableItem>
+										)}
+									</Draggable>
+								})}
+								{provided.placeholder}
+							</div>
 						}}
-					>
-						<Droppable droppableId="droppable">
-							{(provided, snapshot) => {
-								return <div
-									{...provided.droppableProps}
-									ref={provided.innerRef}
-									style={getListStyle(snapshot.isDraggingOver)}
-								>
-									{liveNoteList.map((note, index) => {
-										return <Draggable key={note.id} draggableId={note.id} index={index}>
-											{(provided, snapshot) => (
-												<div
-													ref={provided.innerRef}
-													{...provided.draggableProps}
-													{...provided.dragHandleProps}
-													style={getItemStyle(
-														snapshot.isDragging,
-														provided.draggableProps.style,
-														note.active
-													)}
-													onClick={() => {
-														handleClickLiveNote(note);
-													}}
-													// onContextMenu={(e) => {
-													// 	handleClickLiveNote(note);
-													// 	setCurNote(note);
-
-													// 	e.preventDefault();
-													// 	const clickX = e.clientX;
-													// 	const clickY = e.clientY;
-													// 	const Menu = document.getElementById("Menu");
-													// 	Menu.style.display = "block";
-													// 	Menu.style.position = "absolute";
-													// 	Menu.style.left = `${clickX}px`; Menu.style.top = `${clickY}px`;
-													// }}
-												>
-													{note.title}
-												</div>
-											)}
-										</Draggable>
-									})}
-									{provided.placeholder}
-								</div>
-							}}
-						</Droppable>
-					</DragDropContext>
-				}
-				{/* {liveNoteList.map((note) => {
-					return <Menu.Item key={note.id}
-						onClick={() => {
-							handleClickLiveNote(note);
-						}}
-						onContextMenu={(e) => {
-							handleClickLiveNote(note);
-							setCurNote(note);
-
-							e.preventDefault();
-							const clickX = e.clientX;
-							const clickY = e.clientY;
-							const Menu = document.getElementById("Menu");
-							Menu.style.display = "block";
-							Menu.style.position = "absolute";
-							Menu.style.left = `${clickX}px`; Menu.style.top = `${clickY}px`;
-						}}
-					>
-						<MenuItemContent>
-							<div className='title'>{note.title}</div>
-							<div className='date'>{note.updateTime.format("yyyy/MM/DD")}</div>
-						</MenuItemContent>
-					</Menu.Item>
-				})} */}
+					</Droppable>
+				</DragDropContext>
 			</SubMenu>
-			<SubMenu key="sub2" icon={<DeleteOutlined />} title="Recently Deleted">
+
+			<SubMenu key="sub2" icon={<DeleteOutlined />} title="Trash">
 				{deletedNoteList.map((note) => {
 					return <Menu.Item key={note.id}
 						onClick={() => {
